Add page view tracking, move instantiation function to utils
